requirements.txt
google-adk
pandas
fastapi
uvicorn
python-dotenv
fastapi
uvicorn